pyproject.toml
[tool.poetry]
name = "essence-extractor"
version = "0.1.1"
description = "Unleash the power of content transformation with EssenceExtractor, a dynamic tool that turbocharges your workflow, turning YouTube videos into engaging, readable blog posts in a snap!"
authors = ["Jurik-001"]
readme = "README.md"
packages = [
    { include = "essence_extractor" }
]

[tool.poetry.scripts]
essence-extractor = "essence_extractor.main:args_call"

[tool.poetry.dependencies]
python = "^3.10"
pytube = "^15.0.0"
moviepy = "^1.0.3"
speechrecognition = "^3.10.0"
openai = "^1.2.4"
pytesseract = "^0.3.10"
faiss-cpu = "^1.7.4"
tiktoken = "^0.3.3"
tqdm = "^4.66.1"
pydantic = "^2.5.2"
sentence-transformers = "^2.2.2"
openai-whisper = "^20231117"

[tool.poetry.group.dev.dependencies]
notebook = "^7.0.6"
black = "^23.10.1"
ruff = "^0.1.3"
sphinx = "^7.2.6"
sphinx-rtd-theme = "^2.0.0"
myst-parser = "^2.0.0"
pytest = "^7.4.3"
pytest-cov = "^4.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 90

[tool.ruff]
line-length = 90
select = [
    "E", # Pycodestyle errors: These are coding standard errors found by Pycodestyle (formerly known as Pep8).
    "I", # Isort errors: Issues related to import formatting.
    "D", # Pydocstyle errors: Docstring style issues.
    "COM", # Flake8-commas errors: Issues related to commas (e.g., missing trailing commas).
]
ignore = [
  "D107",  # Missing docstring in __init__
]

[tool.ruff.pydocstyle]
convention = "google"

